Added Users class for working with the database. Upon bot startup, the user is registered in the database.
Localization work has been done. Now all texts support multiple languages with the ability to expand this list.

Added the ability to select a language at startup. This information is saved in the database. Added a function to change the language.

Added a list of services in the form of an inline keyboard.

Navigation has been added for easier interaction with the bot in the future.